Swagger file updated, separated routes into routes/controllers
