--- contacts.family.json
@@ -1,52 +1,0 @@
-[
-  {
-    "_id": {
-      "$oid": "678c1db453b2a3a2f129a175"
-    },
-    "firstName": "Stephanie",
-    "lastName": "Visnaw",
-    "email": "visnaw@gmail.com",
-    "favoriteColor": "Blue",
-    "birthday": "1971-12-11"
-  },
-  {
-    "_id": {
-      "$oid": "678c1db453b2a3a2f129a176"
-    },
-    "firstName": "Dallin",
-    "lastName": "Visnaw",
-    "email": "djvisnaw@gmail.com",
-    "favoriteColor": "Green",
-    "birthday": "1997-06-12"
-  },
-  {
-    "_id": {
-      "$oid": "678c1db453b2a3a2f129a177"
-    },
-    "firstName": "Jayden",
-    "lastName": "Visnaw",
-    "email": "jbvisnaw@gmail.com",
-    "favoriteColor": "Turquoise Blue",
-    "birthday": "2000-03-18"
-  },
-  {
-    "_id": {
-      "$oid": "679547435fe2f4b15fd487f5"
-    },
-    "firstName": "Cameron",
-    "lastName": "Visnaw",
-    "email": "crvisnaw@gmail.com",
-    "favoriteColor": "Purple",
-    "birthday": "2002-05-12"
-  },
-  {
-    "_id": {
-      "$oid": "679572ad83df4bc4ebf52577"
-    },
-    "firstName": "Landon",
-    "lastName": "Visnaw",
-    "email": "lsvisnaw@gmail.com",
-    "favoriteColor": "Purple",
-    "birthday": "2005-04-10"
-  }
-]
