--- requests.rest
@@ -1,72 +1,0 @@
-### Get All Contacts
-GET http://localhost:3000/api/contacts/
-
-### Get Contact by ID
-GET http://localhost:3000/api/contacts/678c1db453b2a3a2f129a177
-
-### Create a new contact
-POST http://localhost:3000/api/contacts
-Content-Type: application/json
-
-{
-    "firstName": "Landon",
-    "lastName": "Visnaw",
-    "email": "lsvisnaw@gmail.com",
-    "favoriteColor": "Blue",
-    "birthday": "2005-04-10"
-}
-
-### Create a new contact to delete and test on
-POST http://localhost:3000/api/contacts
-Content-Type: application/json
-
-{
-    "firstName": "Jimmy Joe III",
-    "lastName": "Johnson",
-    "email": "jjj@gmail.com",
-    "favoriteColor": "Pink",
-    "birthday": "1968-06-13"
-}
-
-### Update a contact
-PUT http://localhost:3000/api/contacts/67958358f3d92e6da5039ce8
-Content-Type: application/json
-
-{
-    "favoriteColor": "Purple"
-}
-
-### Delete a contact
-DELETE http://localhost:3000/api/contacts/67958358f3d92e6da5039ce8
-
-
-### Links for testing live server on render.com
-
-### Get All Contacts
-GET https://cse341-a2-lesson4.onrender.com/api/contacts/
-
-### Get Contact by ID
-GET https://cse341-a2-lesson4.onrender.com/api/contacts/679547435fe2f4b15fd487f5
-
-### Create Contact
-POST https://cse341-a2-lesson4.onrender.com/api/contacts
-Content-Type: application/json
-
-{
-    "firstName": "Jimmy Joe VI",
-    "lastName": "Johnson",
-    "email": "jjjw@gmail.com",
-    "favoriteColor": "Red",
-    "birthday": "2002-05-12"
-}
-
-### Update Contact
-PUT https://cse341-a2-lesson4.onrender.com/api/contacts/<ID>
-Content-Type: application/json
-
-{
-    "favoriteColor": "Pink"
-}
-
-### Delete Contact
-DELETE https://cse341-a2-lesson4.onrender.com/api/contacts/<ID>
